Changed some UI stuff and actually moved non-screen stuff from the app folder.
