--- app/components/index.tsx
@@ -1,13 +1,0 @@
-export { default as Check } from "@/app/components/Check";
-export { default as ContainerGroup } from "@/app/components/ContainerGroup";
-export { default as MatchScoutingHeader } from "@/app/components/MatchScoutingHeader";
-export { default as MatchScoutingNavigation } from "@/app/components/MatchScoutingNavigation";
-export { default as ScoutTeamSelect } from "@/app/components/ScoutTeamSelect";
-export { default as MetricCount } from "@/app/components/MetricCount";
-export { default as MetricLabel } from "@/app/components/MetricLabel";
-export { default as MinusPlusPair } from "@/app/components/MinusPlusPair";
-export { default as QrCodeModal } from "@/app/components/QrCodeModal";
-export { default as ResultsButton } from "@/app/components/ResultsButton";
-export { default as ScoutMatchSelect } from "@/app/components/ScoutMatchSelect";
-export { default as SelectGroup } from "@/app/components/SelectGroup";
-export { default as JsonModal } from "@/app/components/JsonViewModal";
