+++ components/MatchScoutingHeader.tsx
