+++ components/QrCodeModal.tsx
