+++ helpers/fetchFromCosmos.ts
