+++ helpers/refreshMatchScoutingKeys.ts
