+++ helpers/refreshPitScoutingKeys.ts
